rename-variables,attribute-exception,update e2e
