+++ config.yml
@@ -29,7 +29,7 @@ native_service_types:
   - mysql
   - network_drive
   - s3
-  - google_cloud_storage
+  - gcs
 
 # some id
 #connector_id: '1'
@@ -40,4 +40,4 @@ sources:
   dir: connectors.sources.directory:DirectoryDataSource
   mysql: connectors.sources.mysql:MySqlDataSource
   network_drive: connectors.sources.network_drive:NASDataSource
-  google_cloud_storage: connectors.sources.google_cloud_storage:GoogleCloudStorageDataSource
+  gcs: connectors.sources.gcs:GoogleCloudStorageDataSource
