+++ connectors/sources/tests/fixtures/gcs/docker-compose.yml
@@ -1,0 +1,107 @@
+version: '3.9'
+
+services:
+  elasticsearch:
+    image: docker.elastic.co/elasticsearch/elasticsearch:${VERSION}
+    container_name: elasticsearch
+    environment:
+      - cluster.name=docker-cluster
+      - bootstrap.memory_lock=true
+      - ES_JAVA_OPTS=-Xms512m -Xmx512m
+      - ELASTIC_PASSWORD=changeme
+      - xpack.security.enabled=true
+      - xpack.security.authc.api_key.enabled=true
+      - discovery.type=single-node
+      - action.destructive_requires_name=false
+    ulimits:
+      memlock:
+        soft: -1
+        hard: -1
+    volumes:
+      - esdata:/usr/share/elasticsearch/data
+    ports:
+      - 9200:9200
+    networks:
+      - esnet
+
+  mockserver:
+    image: mockserver/mockserver:5.13.2
+    user: root
+    command: -serverPort 443
+    ports:
+      - 443:443
+    environment:
+      MOCKSERVER_WATCH_INITIALIZATION_JSON: "true"
+      MOCKSERVER_PROPERTY_FILE: /gcs/mockserver.properties
+      MOCKSERVER_INITIALIZATION_JSON_PATH: /gcs/mocked_http_responses.json
+      MOCKSERVER_LOG_LEVEL: "ERROR"
+      MOCKSERVER_MAX_EXPECTATIONS: 2000
+      MOCKSERVER_MAX_LOG_ENTRIES: 2000
+    volumes:
+      - type: bind
+        source: .
+        target: /gcs
+    restart: always
+
+  gcs:
+    container_name: gcs
+    image: fsouza/fake-gcs-server
+    ports:
+      - "4443:4443"
+    command: ["-scheme", "http", "-port", "4443"]
+
+  kibana:
+    image: docker.elastic.co/kibana/kibana:${VERSION}
+    ports:
+      - 5601:5601
+    extra_hosts:
+      - "host.docker.internal:host-gateway"
+    depends_on:
+        - "elasticsearch"
+    profiles:
+      - "enterprise-search"
+    environment:
+      ELASTICSEARCH_URL: http://host.docker.internal:9200
+      ELASTICSEARCH_HOSTS: http://host.docker.internal:9200
+      ENTERPRISESEARCH_HOST: http://host.docker.internal:3002
+      ELASTICSEARCH_USERNAME: kibana_system
+      ELASTICSEARCH_PASSWORD: changeme
+    networks:
+      - esnet
+
+  enterprise_search:
+    image: docker.elastic.co/enterprise-search/enterprise-search:${VERSION}
+    profiles:
+      - "enterprise-search"
+    depends_on:
+      - "elasticsearch"
+    environment:
+      - ENT_SEARCH_DEFAULT_PASSWORD=changeme
+      - elasticsearch.username=elastic
+      - elasticsearch.password=changeme
+      - elasticsearch.host=http://host.docker.internal:9200
+      - allow_es_settings_modification=true
+      - kibana.host=http://host.docker.internal:5601
+      - kibana.external_url=http://localhost:5601
+      - secret_management.encryption_keys=["4a2cd3f81d39bf28738c10db0ca782095ffac07279561809eecc722e0c20eb09"]
+      - JAVA_OPTS=-Xms2g -Xmx2g
+      - email.account.enabled=true
+      - email.account.smtp.auth=plain
+      - email.account.smtp.starttls.enable=false
+      - email.account.smtp.host=host.docker.internal
+      - email.account.smtp.port=1025
+      - email.account.email_defaults.from=local@example.com
+      - DEBUG=true
+    ports:
+      - 3002:3002
+    extra_hosts:
+      - "host.docker.internal:host-gateway"
+    networks:
+      - esnet
+
+networks:
+  esnet:
+
+volumes:
+  esdata:
+    driver: local
