+++ connectors/sources/tests/fixtures/gcs/mocked_http_responses.json
@@ -1,0 +1,28 @@
+[
+    {
+        "httpRequest": {
+            "method": "POST",
+            "path": "/token"
+        },
+        "httpResponse": {
+            "body": {
+                "access_token": "XXXXXXStBkRnGyZ2mUYOLgls7QVBxOg82XhBCFo8UIT5gM",
+                "token_type": "Bearer",
+                "expires_in": 3600,
+                "refresh_token": "XXXXXX3SEBX7F2cfrHcqJEa3KoAHYeXES6nmho"
+            },
+            "statusCode": 200
+        }
+    },
+    {
+        "httpRequest": {
+            "method": "GET",
+            "path": "/storage/v1/projects/dummy_project_id/serviceAccount",
+            "secure": true
+        },
+        "httpResponse": {
+            "body": "{'kind': 'storage#serviceAccount', 'email_address': 'service-123456789123@gs-project-accounts.iam.dummy_account.com'}",
+            "statusCode": 200
+        }
+    }
+]
