--- connectors/sources/tests/test_google_cloud_storage.py
@@ -1,509 +1,0 @@
-#
-# Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
-# or more contributor license agreements. Licensed under the Elastic License 2.0;
-# you may not use this file except in compliance with the Elastic License 2.0.
-#
-"""Tests the Google Cloud Storage source class methods.
-"""
-import argparse
-import asyncio
-import json
-import os
-from unittest import mock
-
-import pytest
-from aiogoogle import Aiogoogle
-from aiogoogle.auth.managers import ServiceAccountManager
-from aiogoogle.models import Response, Request
-
-from connectors.source import DataSourceConfiguration
-from connectors.sources.google_cloud_storage import GoogleCloudStorageDataSource
-
-SERVICE_ACCOUNT_CREDENTIAL_PATH = "./path-to-credentials.json"
-API_NAME = "storage"
-API_VERSION = "v1"
-
-
-def get_mocked_source_object():
-    """Creates the mocked Google cloud storage object.
-
-    Returns:
-        GoogleCloudStorageDataSource: Mocked object of the data source class.
-    """
-    connector = argparse.Namespace()
-    connector.configuration = {"credentials_path": SERVICE_ACCOUNT_CREDENTIAL_PATH}
-    dummy_json_data = {"project_id": "dummy123"}
-    with open(
-        file=SERVICE_ACCOUNT_CREDENTIAL_PATH, mode="w", encoding="UTF-8"
-    ) as credentials:
-        json.dump(obj=dummy_json_data, fp=credentials, indent=4)
-
-    mocked_gcs_object = GoogleCloudStorageDataSource(connector=connector)
-    return mocked_gcs_object
-
-
-def test_get_configuration():
-    """Tests the get configurations method of the Google Cloud source class."""
-
-    # Setup
-
-    google_cloud_storage_object = GoogleCloudStorageDataSource
-
-    # Execute
-
-    config = DataSourceConfiguration(
-        config=google_cloud_storage_object.get_default_configuration()
-    )
-
-    # Assert
-
-    assert config["credentials_path"] == "gcs_dummy_credentials.json"
-
-
-@pytest.mark.asyncio
-async def test_ping_for_successful_connection():
-    """Tests the ping functionality for ensuring connection to Google Cloud Storage."""
-
-    # Setup
-
-    expected_response = {
-        "kind": "storage#serviceAccount",
-        "email_address": "serviceaccount@email.com",
-    }
-    mocked_gcs_object = get_mocked_source_object()
-    as_service_account_response = asyncio.Future()
-    as_service_account_response.set_result(expected_response)
-
-    # Execute
-
-    with mock.patch.object(
-        Aiogoogle, "as_service_account", return_value=as_service_account_response
-    ):
-        await mocked_gcs_object.ping()
-
-    # Cleanup
-
-    if os.path.exists(path=SERVICE_ACCOUNT_CREDENTIAL_PATH):
-        os.remove(path=SERVICE_ACCOUNT_CREDENTIAL_PATH)
-
-
-@pytest.mark.asyncio
-async def test_ping_for_failed_connection():
-    """Tests the ping functionality when connection can not be established to Google Cloud Storage."""
-
-    # Setup
-
-    mocked_gcs_object = get_mocked_source_object()
-
-    # Execute
-
-    with pytest.raises(Exception, match="None could not be converted to *"):
-        await mocked_gcs_object.ping()
-
-    # Cleanup
-
-    if os.path.exists(path=SERVICE_ACCOUNT_CREDENTIAL_PATH):
-        os.remove(path=SERVICE_ACCOUNT_CREDENTIAL_PATH)
-
-
-@pytest.mark.parametrize(
-    "previous_documents_list, updated_documents_list",
-    [
-        (
-            [
-                {
-                    "items": [
-                        {
-                            "id": "path_to_blob",
-                            "updated": "2011-10-12T00:00:00Z",
-                            "name": "path_to_blob",
-                        }
-                    ]
-                },
-                {
-                    "kind": "storage#buckets",
-                },
-            ],
-            [
-                {
-                    "_id": "path_to_blob",
-                    "component_count": None,
-                    "content_encoding": None,
-                    "content_language": None,
-                    "created_at": None,
-                    "last_updated": "2011-10-12T00:00:00Z",
-                    "metadata": None,
-                    "name": "path_to_blob",
-                    "size": None,
-                    "storage_class": None,
-                    "_timestamp": "2011-10-12T00:00:00Z",
-                    "type": None,
-                    "url": "https://console.cloud.google.com/storage/browser/_details/None/path_to_blob;tab=live_object?project=dummy123",
-                    "version": None,
-                    "bucket_name": None,
-                }
-            ],
-        )
-    ],
-)
-def test_get_blob_document(previous_documents_list, updated_documents_list):
-    """Tests the function which modifies the fetched blobs and maps the values to keys.
-
-    Args:
-        previous_documents_list (list): List of the blobs documents fetched from Google Cloud Storage.
-        updated_documents_list (list): List of the documents returned by the method.
-    """
-
-    # Setup
-    mocked_gcs_object = get_mocked_source_object()
-
-    # Execute and Assert
-    assert updated_documents_list == list(
-        mocked_gcs_object.get_blob_document(blobs=previous_documents_list[0])
-    )
-
-
-@pytest.mark.asyncio
-async def test_fetch_buckets():
-    """Tests the method which lists the storage buckets available in Google Cloud Storage."""
-
-    # Setup
-    mocked_gcs_object = get_mocked_source_object()
-    expected_response = {
-        "kind": "storage#objects",
-        "items": [
-            {
-                "kind": "storage#object",
-                "id": "bucket_1",
-                "updated": "2011-10-12T00:00:00Z",
-                "name": "bucket_1",
-            }
-        ],
-    }
-    expected_bucket_list = [
-        {
-            "kind": "storage#objects",
-            "items": [
-                {
-                    "kind": "storage#object",
-                    "id": "bucket_1",
-                    "updated": "2011-10-12T00:00:00Z",
-                    "name": "bucket_1",
-                }
-            ],
-        }
-    ]
-    dummy_url = "https://dummy.gcs.com/buckets/b1/objects"
-
-    expected_response_object = Response(
-        status_code=200,
-        url=dummy_url,
-        json=expected_response,
-        req=Request(method="GET", url=dummy_url),
-    )
-
-    # Execute
-    with mock.patch.object(
-        Aiogoogle, "as_service_account", return_value=expected_response_object
-    ):
-        with mock.patch.object(ServiceAccountManager, "refresh"):
-            bucket_list = []
-            async for bucket in mocked_gcs_object.fetch_buckets():
-                bucket_list.append(bucket)
-
-    # Assert
-    assert bucket_list == expected_bucket_list
-
-    # Cleanup
-    if os.path.exists(path=SERVICE_ACCOUNT_CREDENTIAL_PATH):
-        os.remove(path=SERVICE_ACCOUNT_CREDENTIAL_PATH)
-
-
-@pytest.mark.asyncio
-async def test_fetch_blobs():
-    """Tests the method responsible to yield blobs from Google Cloud Storage bucket."""
-
-    # Setup
-    mocked_gcs_object = get_mocked_source_object()
-    expected_bucket_response = {
-        "kind": "storage#objects",
-        "items": [
-            {
-                "kind": "storage#object",
-                "id": "bucket_1",
-                "updated": "2011-10-12T00:01:00Z",
-                "name": "bucket_1",
-            }
-        ],
-    }
-    dummy_blob_response = {
-        "kind": "storage#objects",
-        "items": [
-            {
-                "kind": "storage#object",
-                "id": "bucket_1/blob_1/123123123",
-                "updated": "2011-10-12T00:01:00Z",
-                "name": "blob_1",
-            }
-        ],
-    }
-    dummy_url = "https://dummy.gcs.com/buckets/b1/objects"
-
-    expected_response_object = Response(
-        status_code=200,
-        url=dummy_url,
-        json=dummy_blob_response,
-        req=Request(method="GET", url=dummy_url),
-    )
-
-    # Execute
-    with mock.patch.object(
-        Aiogoogle, "as_service_account", return_value=expected_response_object
-    ):
-        with mock.patch.object(ServiceAccountManager, "refresh"):
-            async for blob_result in mocked_gcs_object.fetch_blobs(
-                buckets=expected_bucket_response
-            ):
-                # Assert
-                assert blob_result == dummy_blob_response
-
-    # Cleanup
-    if os.path.exists(path=SERVICE_ACCOUNT_CREDENTIAL_PATH):
-        os.remove(path=SERVICE_ACCOUNT_CREDENTIAL_PATH)
-
-
-@pytest.mark.asyncio
-async def test_get_docs():
-    """Tests the module responsible to fetch and yield blobs documents from Google Cloud Storage."""
-
-    # Setup
-    mocked_gcs_object = get_mocked_source_object()
-    expected_response = {
-        "kind": "storage#objects",
-        "items": [
-            {
-                "kind": "storage#object",
-                "id": "bucket_1/blob_1/123123123",
-                "updated": "2011-10-12T00:01:00Z",
-                "name": "blob_1",
-            }
-        ],
-    }
-    expected_blob_document = {
-        "_id": "bucket_1/blob_1/123123123",
-        "component_count": None,
-        "content_encoding": None,
-        "content_language": None,
-        "created_at": None,
-        "last_updated": "2011-10-12T00:01:00Z",
-        "metadata": None,
-        "name": "blob_1",
-        "size": None,
-        "storage_class": None,
-        "_timestamp": "2011-10-12T00:01:00Z",
-        "type": None,
-        "url": "https://console.cloud.google.com/storage/browser/_details/None/blob_1;tab=live_object?project=dummy123",
-        "version": None,
-        "bucket_name": None,
-    }
-    dummy_url = "https://dummy.gcs.com/buckets/b1/objects"
-
-    expected_response_object = Response(
-        status_code=200,
-        url=dummy_url,
-        json=expected_response,
-        req=Request(method="GET", url=dummy_url),
-    )
-
-    # Execute and Assert
-    with mock.patch.object(
-        Aiogoogle, "as_service_account", return_value=expected_response_object
-    ):
-        with mock.patch.object(ServiceAccountManager, "refresh"):
-            async for blob_document in mocked_gcs_object.get_docs():
-                assert blob_document[0] == expected_blob_document
-
-    # Cleanup
-    if os.path.exists(path=SERVICE_ACCOUNT_CREDENTIAL_PATH):
-        os.remove(path=SERVICE_ACCOUNT_CREDENTIAL_PATH)
-
-
-@pytest.mark.asyncio
-async def test_get_docs_when_no_buckets_present():
-    """Tests the module responsible to fetch and yield blobs documents from Google Cloud Storage. When
-    Cloud storage does not have any buckets.
-    """
-
-    # Setup
-    mocked_gcs_object = get_mocked_source_object()
-    expected_response = {
-        "kind": "storage#objects",
-    }
-    dummy_url = "https://dummy.gcs.com/buckets/b1/objects"
-
-    expected_response_object = Response(
-        status_code=200,
-        url=dummy_url,
-        json=expected_response,
-        req=Request(method="GET", url=dummy_url),
-    )
-
-    # Execute and Assert
-    with mock.patch.object(
-        Aiogoogle, "as_service_account", return_value=expected_response_object
-    ):
-        with mock.patch.object(ServiceAccountManager, "refresh"):
-            async for blob_document in mocked_gcs_object.get_docs():
-                assert blob_document[0] is None
-
-    # Cleanup
-    if os.path.exists(path=SERVICE_ACCOUNT_CREDENTIAL_PATH):
-        os.remove(path=SERVICE_ACCOUNT_CREDENTIAL_PATH)
-
-
-@pytest.mark.asyncio
-async def test_get_content():
-    """Test the module responsible for fetching the content of the file if it is extractable."""
-
-    # Setup
-    mocked_gcs_object = get_mocked_source_object()
-    blob_document = {
-        "id": "bucket_1/blob_1/123123123",
-        "component_count": None,
-        "content_encoding": None,
-        "content_language": None,
-        "created_at": None,
-        "last_updated": "2011-10-12T00:01:00Z",
-        "metadata": None,
-        "name": "blob_1.txt",
-        "size": "15",
-        "storage_class": None,
-        "_timestamp": "2011-10-12T00:01:00Z",
-        "type": None,
-        "url": "https://console.cloud.google.com/storage/browser/_details/bucket_1/blob_1;tab=live_object?project=dummy123",
-        "version": None,
-        "bucket_name": "bucket_1",
-    }
-    expected_blob_document = {
-        "_id": "bucket_1/blob_1/123123123",
-        "_timestamp": "2011-10-12T00:01:00Z",
-        "text": "storage dummy store",
-    }
-    blob_content_response = "storage dummy store"
-
-    # Execute and Assert
-    with mock.patch.object(
-        Aiogoogle, "as_service_account", return_value=blob_content_response
-    ):
-        google_client = Aiogoogle(
-            service_account_creds=mocked_gcs_object.service_account_credentials
-        )
-        storage_client = await google_client.discover(
-            api_name=API_NAME, api_version=API_VERSION
-        )
-        storage_client.objects = mock.MagicMock()
-        content = await mocked_gcs_object.get_content(
-            blob=blob_document,
-            doit=True,
-        )
-        assert content == expected_blob_document
-
-    # Cleanup
-    if os.path.exists(path=SERVICE_ACCOUNT_CREDENTIAL_PATH):
-        os.remove(path=SERVICE_ACCOUNT_CREDENTIAL_PATH)
-
-
-@pytest.mark.asyncio
-async def test_get_content_when_type_not_supported():
-    """Test the module responsible for fetching the content of the file if it is not extractable or doit is not true."""
-
-    # Setup
-    mocked_gcs_object = get_mocked_source_object()
-    blob_document = {
-        "_id": "bucket_1/blob_1/123123123",
-        "component_count": None,
-        "content_encoding": None,
-        "content_language": None,
-        "created_at": None,
-        "last_updated": "2011-10-12T00:01:00Z",
-        "metadata": None,
-        "name": "blob_1.cc",
-        "size": None,
-        "storage_class": None,
-        "_timestamp": "2011-10-12T00:01:00Z",
-        "type": None,
-        "url": "https://console.cloud.google.com/storage/browser/_details/None/blob_1;tab=live_object?project=dummy123",
-        "version": None,
-        "bucket_name": None,
-    }
-
-    # Execute and Assert
-    google_client = Aiogoogle(
-        service_account_creds=mocked_gcs_object.service_account_credentials
-    )
-    storage_client = await google_client.discover(
-        api_name=API_NAME, api_version=API_VERSION
-    )
-    storage_client.objects = mock.MagicMock()
-    content = await mocked_gcs_object.get_content(
-        blob=blob_document,
-        doit=True,
-    )
-    assert content is None
-
-    content = await mocked_gcs_object.get_content(
-        blob=blob_document,
-    )
-    assert content is None
-
-    # Cleanup
-    if os.path.exists(path=SERVICE_ACCOUNT_CREDENTIAL_PATH):
-        os.remove(path=SERVICE_ACCOUNT_CREDENTIAL_PATH)
-
-
-@pytest.mark.asyncio
-async def test_get_content_when_file_size_is_large():
-    """Test the module responsible for fetching the content of the file if it is not extractable or doit is not true."""
-
-    # Setup
-    mocked_gcs_object = get_mocked_source_object()
-    blob_document = {
-        "_id": "bucket_1/blob_1/123123123",
-        "component_count": None,
-        "content_encoding": None,
-        "content_language": None,
-        "created_at": None,
-        "last_updated": "2011-10-12T00:01:00Z",
-        "metadata": None,
-        "name": "blob_1.txt",
-        "size": 10000000000000,
-        "storage_class": None,
-        "_timestamp": "2011-10-12T00:01:00Z",
-        "type": None,
-        "url": "https://console.cloud.google.com/storage/browser/_details/None/blob_1;tab=live_object?project=dummy123",
-        "version": None,
-        "bucket_name": None,
-    }
-
-    # Execute and Assert
-    google_client = Aiogoogle(
-        service_account_creds=mocked_gcs_object.service_account_credentials
-    )
-    storage_client = await google_client.discover(
-        api_name=API_NAME, api_version=API_VERSION
-    )
-    storage_client.objects = mock.MagicMock()
-    content = await mocked_gcs_object.get_content(
-        blob=blob_document,
-        doit=True,
-    )
-    assert content is None
-
-    content = await mocked_gcs_object.get_content(
-        blob=blob_document,
-    )
-    assert content is None
-
-    # Cleanup
-    if os.path.exists(path=SERVICE_ACCOUNT_CREDENTIAL_PATH):
-        os.remove(path=SERVICE_ACCOUNT_CREDENTIAL_PATH)
